+++ lib/shrine/plugins/presign_endpoint.rb
@@ -220,9 +220,11 @@ class Shrine
             data = storage.presign(location, options)
           end
 
-          if data.is_a?(Hash)
-            { fields: {}, headers: {} }.merge(data)
+          if data.respond_to?(:to_h)
+            { fields: {}, headers: {} }.merge(data.to_h)
           else
+            Shrine.deprecation("Returning a custom object in Storage#presign is deprecated, presign_endpoint will not support it in Shrine 3. Storage#presign should return a Hash instead.")
+
             url     = data.url
             fields  = data.fields
             headers = data.headers if data.respond_to?(:headers)